Refactoring of /videos page and VideoTable component
